Give ourselves a new stack every time we run
When (built-in) functions are invoked they always store their
return values upon the stack.  Usually this is OK because the
return value will be used for something, and thus popped-off.

However people will rarely test to result of calling `print`,
which means the stack will grow one element every time that
function is invoked.

Since we allow our `Run` method to be reused on the same
script, and with the same environment, it is possible we'll
have a stack growing to an essentially infinite size if a
script is constantly reused.

To avoid that explicitly give ourself a new stack every time
we run a script.

This closes #86.