Simplify the replacement of opcode arguments.
This is part of #90, as I see now how simpler other
implementations are.  We don't need to bother with
anything except the opcode argument in our implementation.

While it is true that some steps of our optimizer change
instructions in our main compilere we only update the target
of our two branching instructions `OpJump` and `OpJumpIfFalse`
for the handling of `if`, `while`, and ternary statements.