Add a void-object, and use it.
The void-object allows a built-in function to NOT return a result,
which means that we won't needlessly pad the stack with an extra
value which isn't used.

This will be very helpful when people extend the engine with
their own primitives, and also for the built in `printf` and `print`
functions.

This closes #106.