Compile functions.
This is a sleazy hack.  Unlike the WACiG we're doing this in
a naive and hacky way - instead of using proper scopes we
can cheat and just reset our instruction-list to an empty set.

This is going to work out just fine, because we're going to
essentially execute any function in a new scope, in a new
VM, so the important thing is that our bytecode-references
(and their fixups) start from zero.