Fixed leaking scopes!
This commit makes everything work the way that I wanted, there
was a bug in the sense that a new scope would create a local
variable - which would then be thrown away when the scope
was exited.

This lead to:

    function foo(name)
    {
       // name is valid here ..

      foreach char in name {
         // changes to name here failed
         // since there was a NEW copy
         // as we enter a new scope in the
         // foreach

The solution to this is to allow existing variables to be
updated in the appropriate-scope, not the most recent.

That took some finding.