Remove functions that are never called.
This commit removes functions which are not called, in an attempt
at optimization.  Honestly there will be no runtime difference,
but it feels like the right thing to do.

Given this input:

```
function foo() {
  foreach char in "Steve" {
   printf( "Char: %s\n", char );
  }
}

// foo();
return false;
```

Our generated bytecode is `OpFalse; OpReturn`, and there are no
references to the function `foo` at all.

This closes #140.

However the constants are still present, which is tracked in #141:

```
$ go build . && ./evalfilter bytecode ./t.in
Bytecode:
  0000	       OpFalse
  0001	      OpReturn

Constant Pool:
  0000 Type:STRING Value:"Steve"
  0001 Type:STRING Value:""
  0002 Type:STRING Value:"char"
  0003 Type:STRING Value:"Char: %s\n"
  0004 Type:STRING Value:"printf"
```