Second attempt at nested fields
This is a work in-progress which moves towards a simpler solution
for nested field access.

The realization that we need to only handle the case of maps, and
allow parsing them to our internal hash-objects.

If we have hash-objects parsed correctly then we can add the "."
operator as a synonym, so these two lines of code are identical:

       foo["bar"]
       foo.bar

This commit makes this change, the harder one of parsing the map
is left as a followup.